move tab methods to trait
