+++ app/Http/Livewire/Delegates/Tabs.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace App\Http\Livewire\Delegates;
 
-use App\Livewire\SupportBrowserHistoryWrapper;
+use App\Http\Livewire\Concerns\HasTabs;
 use Illuminate\Contracts\View\View;
-use Illuminate\Support\Arr;
-use Illuminate\Support\Str;
 use Livewire\Component;
 
 final class Tabs extends Component
 {
+    use HasTabs;
+
     public string $view = 'delegates';
 
     public ?string $previousView = 'delegates';
@@ -21,48 +21,20 @@ final class Tabs extends Component
     public array $savedQueryData = [];
 
     public array $alreadyLoadedViews = [
-        'delegates'     => false,
-        'missed-blocks' => false,
-        'recent-votes'  => false,
+        'delegates' => false,
     ];
-
-    // /** @var mixed */
-    // protected $listeners = [
-    //     'showWalletView',
-    // ];
-
-    public function __get(mixed $property): mixed
-    {
-        $value = Arr::get($this->tabQueryData[$this->view], $property);
-        if ($value !== null) {
-            return $value;
-        }
-
-        $value = Arr::get($this->tabQueryData[$this->previousView], $property);
-        if ($value !== null) {
-            return $value;
-        }
-
-        return parent::__get($property);
-    }
-
-    public function __set(string $property, mixed $value): void
-    {
-        if (Arr::has($this->tabQueryData[$this->view], $property)) {
-            $this->tabQueryData[$this->view][$property] = $value;
-        }
-
-        if (Arr::has($this->tabQueryData[$this->previousView], $property)) {
-            $this->tabQueryData[$this->previousView][$property] = $value;
-        }
-    }
 
     public function queryString(): array
     {
+        $perPage = intval(config('arkscan.pagination.per_page'));
+        if ($this->view === 'delegates') {
+            $perPage = Delegates::defaultPerPage();
+        }
+
         $params = [
             'view'    => ['except' => 'delegates'],
             'page'    => ['except' => 1],
-            'perPage' => ['except' => intval(config('arkscan.pagination.per_page'))],
+            'perPage' => ['except' => $perPage],
         ];
 
         // We need to pass in the filters for previous view so we can hide it from the URL
@@ -73,14 +45,7 @@ final class Tabs extends Component
         return [
             ...$params,
 
-            // TODO: Filters - https://app.clickup.com/t/861n4ydmh
-            // Filters
-            // 'outgoing'      => ['except' => true],
-            // 'incoming'      => ['except' => true],
-            // 'transfers'     => ['except' => true],
-            // 'votes'         => ['except' => true],
-            // 'multipayments' => ['except' => true],
-            // 'others'        => ['except' => true],
+            // TODO: Filters - https://app.clickup.com/t/861n4ydmh - see WalletTables
         ];
     }
 
@@ -89,26 +54,10 @@ final class Tabs extends Component
         if ($this->tabQueryData === []) {
             $this->tabQueryData = [
                 'delegates' => [
-                    'page'          => 1,
-                    'perPage'       => 51, //WalletTransactionTable::defaultPerPage(),
-
-                    // TODO: Filters - https://app.clickup.com/t/861n4ydmh
-                    'outgoing'      => true,
-                    'incoming'      => true,
-                    'transfers'     => true,
-                    'votes'         => true,
-                    'multipayments' => true,
-                    'others'        => true,
-                ],
-
-                'missed-blocks' => [
                     'page'    => 1,
-                    'perPage' => 1, //WalletBlockTable::defaultPerPage(),
-                ],
+                    'perPage' => Delegates::defaultPerPage(),
 
-                'recent-votes' => [
-                    'page'    => 1,
-                    'perPage' => 1, //WalletVoterTable::defaultPerPage(),
+                    // TODO: Filters - https://app.clickup.com/t/861n4ydmh - see WalletTables
                 ],
             ];
         }
@@ -117,66 +66,5 @@ final class Tabs extends Component
     public function render(): View
     {
         return view('livewire.delegates.tabs');
-    }
-
-    // public function showWalletView(string $view): void
-    // {
-    //     $this->view = $view;
-    // }
-
-    public function triggerViewIsReady(?string $view = null): void
-    {
-        if ($view === null) {
-            $view = $this->view;
-        }
-
-        if (! array_key_exists($view, $this->alreadyLoadedViews)) {
-            return;
-        }
-
-        if ($this->alreadyLoadedViews[$view] === true) {
-            return;
-        }
-
-        $this->emit('set'.Str::studly($view).'Ready');
-
-        $this->alreadyLoadedViews[$view] = true;
-    }
-
-    public function updatingView(string $newView): void
-    {
-        if ($newView === $this->view) {
-            return;
-        }
-
-        $this->previousView = $this->view;
-
-        SupportBrowserHistoryWrapper::init()->mergeRequestQueryStringWithComponent($this);
-
-        $this->savedQueryData[$this->view] = $this->tabQueryData[$this->view];
-
-        // Reset the querystring data on view change to clear the URL
-        $queryStringData = $this->queryString();
-        foreach ($this->tabQueryData[$this->view] as $key => $value) {
-            // @phpstan-ignore-next-line
-            $this->{$key} = $queryStringData[$key]['except'];
-        }
-
-        $this->triggerViewIsReady($newView);
-    }
-
-    /**
-     * Apply existing view data to query string.
-     *
-     * @return void
-     */
-    public function updatedView(): void
-    {
-        if (array_key_exists($this->view, $this->savedQueryData)) {
-            foreach ($this->savedQueryData[$this->view] as $key => $value) {
-                // @phpstan-ignore-next-line
-                $this->{$key} = $value;
-            }
-        }
     }
 }
